fix some bug

+++ app-web/src/main/java/com/anhtuan/store/dto/response/DiscountResponseDto.java
@@ -5,6 +5,6 @@ import lombok.Data;
 @Data
 public class DiscountResponseDto {
     private Integer id;
-    private Double percent;
+    private Integer percent;
     private String description;
 }
